+++ TicketCentral/Pages/Tickets/Delete.cshtml.cs
@@ -11,36 +11,30 @@ namespace TicketCentral.Pages.Tickets
 {
     public class DeleteModel : PageModel
     {
-        private readonly TicketCentral.Models.TicketContext _context;
+        private readonly TicketCentral.Models.BookingContext _context;
 
-        public DeleteModel(TicketCentral.Models.TicketContext context)
+        public DeleteModel(TicketCentral.Models.BookingContext context)
         {
             _context = context;
         }
 
         [BindProperty]
         public Ticket Ticket { get; set; }
-        public string ErrorMessage { get; set; }
 
-        public async Task<IActionResult> OnGetAsync(int? id, bool? saveChangesError = false)
+        public async Task<IActionResult> OnGetAsync(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            Ticket = await _context.Ticket.AsNoTracking()
+            Ticket = await _context.Ticket
                 .Include(t => t.Customer)
                 .Include(t => t.VenueBooking).FirstOrDefaultAsync(m => m.TicketID == id);
 
             if (Ticket == null)
             {
                 return NotFound();
-            }
-
-            if (saveChangesError.GetValueOrDefault())
-            {
-                ErrorMessage = "Delete failed. Try again";
             }
             return Page();
         }
@@ -52,19 +46,15 @@ namespace TicketCentral.Pages.Tickets
                 return NotFound();
             }
 
-            Ticket = await _context.Ticket.AsNoTracking()
-                .FirstOrDefaultAsync(t => t.TicketID == id);
+            Ticket = await _context.Ticket.FindAsync(id);
 
-            try
+            if (Ticket != null)
             {
                 _context.Ticket.Remove(Ticket);
                 await _context.SaveChangesAsync();
-                return RedirectToPage("./Index");
             }
-            catch (DbUpdateException)
-            {
-                return RedirectToAction(".Delete", new { id, saveChangesError = true });
-            }
+
+            return RedirectToPage("./Index");
         }
     }
 }
